# artifacts/api-server/src/models/Notification.ts
import mongoose from "mongoose";

const notificationSchema = new mongoose.Schema(
  {
    type: {
      type: String,
      enum: ["renewal_alert", "new_member", "member_deleted", "manual_reminder"],
      required: true,
    },
    title: { type: String, required: true },
    message: { type: String, required: true },
    memberId: { type: mongoose.Schema.Types.ObjectId, ref: "Member", default: null },
    isRead: { type: Boolean, default: false },
  },
  { timestamps: true }
);

export const Notification = mongoose.model("Notification", notificationSchema);