# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Gym CRM API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /members:
    get:
      operationId: getMembers
      tags: [members]
      summary: Get all members
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [Active, "Expiring Soon", Expired]
        - name: packageType
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Members list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MembersListResponse"
    post:
      operationId: createMember
      tags: [members]
      summary: Create a member
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberResponse"
  /members/{id}:
    get:
      operationId: getMember
      tags: [members]
      summary: Get member by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberResponse"
    put:
      operationId: updateMember
      tags: [members]
      summary: Update member
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberResponse"
    delete:
      operationId: deleteMember
      tags: [members]
      summary: Delete member
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /members/{id}/remind:
    post:
      operationId: sendMemberReminder
      tags: [members]
      summary: Send reminder to member
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reminder sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /notifications:
    get:
      operationId: getNotifications
      tags: [notifications]
      summary: Get all notifications
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Notifications list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationsListResponse"
  /notifications/unread-count:
    get:
      operationId: getUnreadCount
      tags: [notifications]
      summary: Get unread notification count
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Unread count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnreadCountResponse"
  /notifications/{id}/read:
    put:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /notifications/read-all:
    put:
      operationId: markAllNotificationsRead
      tags: [notifications]
      summary: Mark all notifications as read
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /notifications/clear-all:
    delete:
      operationId: clearAllNotifications
      tags: [notifications]
      summary: Clear all notifications
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All cleared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get settings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettingsResponse"
    put:
      operationId: updateSettings
      tags: [settings]
      summary: Update settings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingsInput"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettingsResponse"
  /settings/test-email:
    post:
      operationId: sendTestEmail
      tags: [settings]
      summary: Send test email
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Test email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    Member:
      type: object
      properties:
        _id:
          type: string
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        profilePhoto:
          type: string
        packageType:
          type: string
          enum: [Daily, Monthly, "6 Months", "1 Year"]
        packageStartDate:
          type: string
        packageEndDate:
          type: string
        totalFeesPaid:
          type: number
        notes:
          type: string
        lastReminderSent:
          type: string
        status:
          type: string
        daysRemaining:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - _id
        - fullName
        - email
        - phone
        - packageType
        - packageStartDate
        - packageEndDate
    MemberInput:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        profilePhoto:
          type: string
        packageType:
          type: string
          enum: [Daily, Monthly, "6 Months", "1 Year"]
        packageStartDate:
          type: string
        totalFeesPaid:
          type: number
        notes:
          type: string
      required:
        - fullName
        - email
        - phone
        - packageType
        - packageStartDate
    MemberResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/Member"
        message:
          type: string
      required:
        - success
    MembersListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            members:
              type: array
              items:
                $ref: "#/components/schemas/Member"
            total:
              type: integer
            page:
              type: integer
            totalPages:
              type: integer
          required:
            - members
            - total
      required:
        - success
    Notification:
      type: object
      properties:
        _id:
          type: string
        type:
          type: string
          enum: [renewal_alert, new_member, member_deleted, manual_reminder]
        title:
          type: string
        message:
          type: string
        memberId:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
      required:
        - _id
        - type
        - title
        - message
        - isRead
        - createdAt
    NotificationsListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
      required:
        - success
    UnreadCountResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            count:
              type: integer
          required:
            - count
      required:
        - success
...
    Settings:
      type: object
      properties:
        gymName:
          type: string
        gymAddress:
          type: string
        gymPhone:
          type: string
        gymLogo:
          type: string
        emailUser:
          type: string
        adminFullName:
          type: string
        adminEmail:
          type: string
      required:
        - gymName
    SettingsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/Settings"
      required:
        - success
    SettingsInput:
      type: object
      properties:
        gymName:
          type: string
        gymAddress:
          type: string
        gymPhone:
          type: string
        gymLogo:
          type: string
        emailUser:
          type: string
        emailPass:
          type: string
        adminFullName:
          type: string
        currentPassword:
          type: string
        newPassword:
          type: string