# lib/api-zod/src/generated/types/membersListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gym CRM API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MembersListResponseData } from "./membersListResponseData";

export interface MembersListResponse {
  success: boolean;
  data?: MembersListResponseData;
}